tg_bot/requirements.txt
elasticsearch>=7.0.0,<8.0.0
python-dateutil==2.8.1
python-telegram-bot==12.8
